+++ apps/chatbot-server/ingest/sources/github-leafygreen-ui.ts
@@ -4,6 +4,25 @@ import {
   makeGitDataSource,
 } from 'mongodb-rag-core/dataSources';
 
+export const leafygreenGithubSourceConstructor = async () => {
+  return await makeGitDataSource({
+    name: 'leafygreen-ui',
+    repoUri: 'https://github.com/mongodb/leafygreen-ui.git',
+    repoOptions: {
+      '--depth': 1,
+      '--branch': 'main',
+    },
+    metadata: {
+      productName: 'LeafyGreen UI',
+      version: '1.0.0',
+      tags: ['leafygreen', 'docs'],
+    },
+    filter: (path: string) => path.endsWith('.md') || path.includes('types'),
+    handlePage: async (path, content) =>
+      await handleHtmlDocument(path, content, htmlParserOptions),
+  });
+};
+
 const removeElements = (domDoc: Document) => [
   ...Array.from(domDoc.querySelectorAll('head')),
   ...Array.from(domDoc.querySelectorAll('script')),
@@ -37,23 +56,4 @@ const htmlParserOptions: Omit<HandleHtmlPageFuncOptions, 'sourceName'> = {
   removeElements,
   extractTitle,
   extractMetadata,
-};
-
-export const leafygreenGithubSourceConstructor = async () => {
-  return await makeGitDataSource({
-    name: 'leafygreen-ui',
-    repoUri: 'https://github.com/mongodb/leafygreen-ui.git',
-    repoOptions: {
-      '--depth': 1,
-      '--branch': 'main',
-    },
-    metadata: {
-      productName: 'LeafyGreen UI',
-      version: '1.0.0',
-      tags: ['leafygreen', 'docs'],
-    },
-    filter: (path: string) => path.endsWith('.md') || path.includes('types'),
-    handlePage: async (path, content) =>
-      await handleHtmlDocument(path, content, htmlParserOptions),
-  });
 };
